initial work to offload feature handling to cligj
