+++ mapboxcli/errors.py
@@ -1,0 +1,4 @@
+import click
+
+class MapboxCLIException(click.ClickException):
+    pass
